TIL/03.25/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
Pillow
ultralytics
scikit-learn
torch